react/kiez-search-profile: add initial dashboard
